Replace decimal with double.

+++ PierresBakery/Models/Bread.cs
@@ -3,13 +3,13 @@ namespace PierresBakery.Models
   public class Bread : BakedGood
   {
     public static int BreadCount { get; set; }
-    public static decimal PriceForOne { get; private set; }
-    public static decimal PriceForDeal { get; private set; }
+    public static double PriceForOne { get; private set; }
+    public static double PriceForDeal { get; private set; }
     public static int NumberOfBakedGoodsForDeal { get; private set; }
     public static int NumberOfBakedGoodsFreeFromDeal { get; private set; }
     static Bread()
     {
-      PriceForOne = 5.00m;
+      PriceForOne = 5.00;
       PriceForDeal = 2 * PriceForOne;
       NumberOfBakedGoodsForDeal = 2;
       NumberOfBakedGoodsFreeFromDeal = 1;
@@ -19,7 +19,7 @@ namespace PierresBakery.Models
 
     }
 
-    public static decimal CalculateCostForBakedGoods(int numberOfBakedGoods)
+    public static double CalculateCostForBakedGoods(int numberOfBakedGoods)
     {
       int numberOfBakedGoodsDeals = 0;
       int numberOfRegularPriceBakeGoods = 0;
@@ -33,8 +33,8 @@ namespace PierresBakery.Models
         numberOfRegularPriceBakeGoods = numberOfBakedGoods % (NumberOfBakedGoodsForDeal + NumberOfBakedGoodsFreeFromDeal);
       }
 
-      decimal costOfBakedGoodsDeals = numberOfBakedGoodsDeals * PriceForDeal;
-      decimal costOfRegularPricedBakedGood = numberOfRegularPriceBakeGoods * PriceForOne;
+      double costOfBakedGoodsDeals = numberOfBakedGoodsDeals * PriceForDeal;
+      double costOfRegularPricedBakedGood = numberOfRegularPriceBakeGoods * PriceForOne;
       return costOfBakedGoodsDeals + costOfRegularPricedBakedGood;
     }
   }
