+++ PierresBakery/Models/Pastry.cs
@@ -2,14 +2,14 @@ namespace PierresBakery.Models
 {
   public class Pastry : BakedGood
   {
-    public static decimal PriceForOne { get; private set; }
-    public static decimal PriceForDeal { get; private set; }
+    public static double PriceForOne { get; private set; }
+    public static double PriceForDeal { get; private set; }
     public static int NumberOfBakedGoodsForDeal { get; private set; }
     public static int NumberOfBakedGoodsFreeFromDeal { get; private set; }
     static Pastry()
     {
-      PriceForOne = 2.00m;
-      PriceForDeal = 2.5m * PriceForOne;
+      PriceForOne = 2.00;
+      PriceForDeal = 2.5 * PriceForOne;
       NumberOfBakedGoodsForDeal = 3;
       NumberOfBakedGoodsFreeFromDeal = 0;
     }
@@ -17,7 +17,7 @@ namespace PierresBakery.Models
     {
 
     }
-    public static decimal CalculateCostForBakedGoods(int numberOfBakedGoods)
+    public static double CalculateCostForBakedGoods(int numberOfBakedGoods)
     {
       int numberOfBakedGoodsDeals = 0;
       int numberOfRegularPriceBakeGoods = 0;
@@ -31,8 +31,8 @@ namespace PierresBakery.Models
         numberOfRegularPriceBakeGoods = numberOfBakedGoods % (NumberOfBakedGoodsForDeal + NumberOfBakedGoodsFreeFromDeal);
       }
 
-      decimal costOfBakedGoodsDeals = numberOfBakedGoodsDeals * PriceForDeal;
-      decimal costOfRegularPricedBakedGood = numberOfRegularPriceBakeGoods * PriceForOne;
+      double costOfBakedGoodsDeals = numberOfBakedGoodsDeals * PriceForDeal;
+      double costOfRegularPricedBakedGood = numberOfRegularPriceBakeGoods * PriceForOne;
       return costOfBakedGoodsDeals + costOfRegularPricedBakedGood;
     }
   }
